build: Add options to select sanitizers in configure.py
We are going to add ThreadSanitizer that can't be used together with
AddressSanitizer. We want to choose which sanitizers to use.

There are options to use sanitizers:

Use ASAN and UBSAN for Debug and Sanitize build types:
./configure.py

Use specified sanitizers for Debug and Sanitize build types:
./configure.py --sanitizer address --sanitizer undefined_behavior

Do not use sanitizers for any build type:
./configure.py --no-sanitizers

Enabling sanitizers is consistently with Seastar_SANITIZE option.

Specified sanitizers are passed to Seastar_SANITIZERS list.